+++ core/banking/gov_transfers.go
@@ -92,6 +92,12 @@ func (e *Engine) distributeRecurringGovernanceTransfers(ctx context.Context) {
 			continue
 		}
 
+		if gTransfer.Config.RecurringTransferConfig.DispatchStrategy != nil && gTransfer.Config.RecurringTransferConfig.DispatchStrategy.TransferInterval != nil &&
+			(e.currentEpoch-gTransfer.Config.RecurringTransferConfig.StartEpoch+1) < uint64(*gTransfer.Config.RecurringTransferConfig.DispatchStrategy.TransferInterval) &&
+			(e.currentEpoch-gTransfer.Config.RecurringTransferConfig.StartEpoch+1)%uint64(*gTransfer.Config.RecurringTransferConfig.DispatchStrategy.TransferInterval) != 0 {
+			continue
+		}
+
 		amount, err := e.processGovernanceTransfer(ctx, gTransfer)
 		e.log.Info("processed transfer", logging.String("amount", amount.String()))