+++ core/rewards/mocks/mocks.go
@@ -83,17 +83,17 @@ func (mr *MockMarketActivityTrackerMockRecorder) GetAllMarketIDs() *gomock.Call 
 }
 
 // GetLastEpochTakeFees mocks base method.
-func (m *MockMarketActivityTracker) GetLastEpochTakeFees(arg0 string, arg1 []string) map[string]*num.Uint {
+func (m *MockMarketActivityTracker) GetLastEpochTakeFees(arg0 string, arg1 []string, arg2 int32) map[string]*num.Uint {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "GetLastEpochTakeFees", arg0, arg1)
+	ret := m.ctrl.Call(m, "GetLastEpochTakeFees", arg0, arg1, arg2)
 	ret0, _ := ret[0].(map[string]*num.Uint)
 	return ret0
 }
 
 // GetLastEpochTakeFees indicates an expected call of GetLastEpochTakeFees.
-func (mr *MockMarketActivityTrackerMockRecorder) GetLastEpochTakeFees(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockMarketActivityTrackerMockRecorder) GetLastEpochTakeFees(arg0, arg1, arg2 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetLastEpochTakeFees", reflect.TypeOf((*MockMarketActivityTracker)(nil).GetLastEpochTakeFees), arg0, arg1)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetLastEpochTakeFees", reflect.TypeOf((*MockMarketActivityTracker)(nil).GetLastEpochTakeFees), arg0, arg1, arg2)
 }
 
 // GetProposer mocks base method.